+++ CHANGELOG.md
@@ -8,5 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Timeout parameter to ReadRawAccelerationsFromFifo and ReadAccelerationsFromFifo.
 
+### Fixed
+- SelfTest not restoring the acceleration range.
+- SelfTest not clearing the FIFO.
+
 ## [1.0.0] - 2024-08-27
 Initial release.